configure web sample
