+++ source/NHibernate.AspNet.Web.Specs/app.config
@@ -4,12 +4,15 @@
     <section name="hibernate-configuration" type="NHibernate.Cfg.ConfigurationSectionHandler, NHibernate" />
   </configSections>
   <hibernate-configuration xmlns="urn:nhibernate-configuration-2.2">
+    <bytecode-provider type="null"/><!-- Important under Medium Trust -->
     <session-factory>
+      <property name="connection.provider">NHibernate.Connection.DriverConnectionProvider, NHibernate</property>
       <property name="dialect">NHibernate.Dialect.SQLiteDialect</property>
       <property name="connection.driver_class">NHibernate.Driver.SQLite20Driver</property>
       <property name="connection.connection_string_name">DefaultConnection</property>
       <property name="connection.release_mode">on_close</property>
       <property name="show_sql">true</property>
+      <property name="current_session_context_class">thread_static</property>
     </session-factory>
   </hibernate-configuration>
   <connectionStrings>
