+++ source/NHibernate.AspNet.Web.Specs/packages.config
@@ -11,7 +11,6 @@
   <package id="Microsoft.AspNet.WebPages" version="3.2.3" targetFramework="net451" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net451" />
   <package id="MilesiBastos.SharpArch.Domain" version="4.0.3000" targetFramework="net451" />
-  <package id="MilesiBastos.SharpArch.NHibernate" version="4.0.3000" targetFramework="net451" />
   <package id="Moq" version="4.2.1502.0911" targetFramework="net451" />
   <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net451" />
   <package id="NHibernate" version="4.0.3.4000" targetFramework="net451" />
