+++ source/NHibernate.AspNet.Web/Startup.cs
@@ -1,6 +1,11 @@
-﻿using Microsoft.Owin;
+﻿using System;
+using System.IO;
+using Microsoft.Owin;
+using NHibernate.AspNet.Identity.Helpers;
+using NHibernate.AspNet.Web.Models;
+using NHibernate.Cfg;
+using NHibernate.Tool.hbm2ddl;
 using Owin;
-using SharpArch.NHibernate.Web.Mvc;
 
 [assembly: OwinStartupAttribute(typeof(NHibernate.AspNet.Web.Startup))]
 namespace NHibernate.AspNet.Web
@@ -10,13 +15,33 @@ namespace NHibernate.AspNet.Web
         public void Configuration(IAppBuilder app)
         {
             ConfigureAuth(app);
-            ConfigureData();
+
+            var internalTypes = new[] {
+                typeof(ApplicationUser)
+            };
+
+            var mapping = MappingHelper.GetIdentityMappings(internalTypes);
+
+            var config = new Configuration().Configure();
+            config.AddDeserializedMapping(mapping, null);
+            BuildSchema(config);
+
+            var builder = new SessionFactoryBuilder();
+            builder.RegisterToBuild(config);
+            // Register a second config for multi data bases
+            // builder.RegisterToBuild(secondconfig); 
+            builder.Build(app);
         }
 
-        private static void ConfigureData()
+        private static void BuildSchema(Configuration config)
         {
-            var storage = new WebSessionStorage(System.Web.HttpContext.Current.ApplicationInstance);
-            DataConfig.Configure(storage);
+            var path = Path.Combine(AppDomain.CurrentDomain.GetData("DataDirectory").ToString(), @"schema.sql");
+
+            // this NHibernate tool takes a configuration (with mapping info in)
+            // and exports a database schema from it
+            new SchemaExport(config)
+                .SetOutputFile(path)
+                .Create(true, true /* DROP AND CREATE SCHEMA */);
         }
     }
 }
