+++ source/NHibernate.AspNet.Web/Web.config
@@ -5,11 +5,13 @@
   </configSections>
   <hibernate-configuration xmlns="urn:nhibernate-configuration-2.2">
     <session-factory>
+      <property name="connection.provider">NHibernate.Connection.DriverConnectionProvider, NHibernate</property>
       <property name="dialect">NHibernate.Dialect.SQLiteDialect</property>
       <property name="connection.driver_class">NHibernate.Driver.SQLite20Driver</property>
       <property name="connection.connection_string_name">DefaultConnection</property>
       <property name="connection.release_mode">on_close</property>
       <property name="show_sql">true</property>
+      <property name="current_session_context_class">web</property>
     </session-factory>
   </hibernate-configuration>
   <connectionStrings>
