+++ source/NHibernate.AspNet.Web/packages.config
@@ -25,7 +25,6 @@
   <package id="Microsoft.Owin.Security.Twitter" version="3.0.1" targetFramework="net451" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net451" />
   <package id="MilesiBastos.SharpArch.Domain" version="4.0.3000" targetFramework="net451" />
-  <package id="MilesiBastos.SharpArch.NHibernate" version="4.0.3000" targetFramework="net451" />
   <package id="Modernizr" version="2.8.3" targetFramework="net451" />
   <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net451" />
   <package id="NHibernate" version="4.0.3.4000" targetFramework="net451" />
